.github/requirements.txt
rapyuta-io @ https://storage.googleapis.com/rio-sdk-python/rapyuta_io-0.18.1-py2-none-any.whl
PyYAML==5.4.1